+++ src/utils/event-utils.ts
@@ -1,4 +1,4 @@
-import { _getQueryParam } from './request-utils'
+import { _getQueryParam, convertToURL } from './request-utils'
 import { _isNull, _isUndefined } from './type-utils'
 import { Properties } from '../types'
 import Config from '../config'
@@ -16,6 +16,7 @@ function isSafari(userAgent: string): boolean {
 
 export const _info = {
     campaignParams: function (customParams?: string[]): Record<string, any> {
+        // Should be kept in sync with https://github.com/PostHog/posthog/blob/master/plugin-server/src/utils/db/utils.ts#L60
         const campaign_keywords = [
             'utm_source',
             'utm_medium',
@@ -23,6 +24,8 @@ export const _info = {
             'utm_content',
             'utm_term',
             'gclid',
+            'gbraid',
+            'wbraid',
             'fbclid',
             'msclkid',
         ].concat(customParams || [])
@@ -256,9 +259,7 @@ export const _info = {
         if (!document?.referrer) {
             return '$direct'
         }
-        const parser = document.createElement('a') // Unfortunately we cannot use new URL due to IE11
-        parser.href = document.referrer
-        return parser.host
+        return convertToURL(document.referrer)?.host || '$direct'
     },
 
     properties: function (): Properties {