+++ src/utils/index.ts
@@ -181,7 +181,7 @@ export const _safewrap = function <F extends (...args: any[]) => any = (...args:
             // @ts-ignore
             return f.apply(this, args)
         } catch (e) {
-            logger.critical('Implementation error. Please turn on debug and contact support@posthog.com.')
+            logger.critical('Implementation error. Please turn on debug mode and open a ticket on https://app.posthog.com/home#panel=support%3Asupport%3A.')
             logger.critical(e)
         }
     } as F
@@ -486,4 +486,8 @@ export function isCrossDomainCookie(documentLocation: Location | undefined) {
     // but it's good enough for ensuring we only match herokuapp.com when it is the TLD
     // for the hostname
     return hostname.split('.').slice(-2).join('.') !== 'herokuapp.com'
+}
+
+export function isDistinctIdStringLike(value: string): boolean {
+    return ['distinct_id', 'distinctid'].includes(value.toLowerCase())
 }